template_app/app/requirements.txt
streamlit==1.41.0
openai==1.57.2
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
html2docx==1.6.0
xhtml2pdf==0.2.16
reportlab==4.2.5
